Solution to Check Even or Odd

+++ Check Even or Odd@@6788dd63490637cfc44e1b8d/code.c
@@ -2,7 +2,7 @@
 
 int main() 
 {
-    float num;
+    int num;
 
     if (num%2==0){
         printf("Even");
